Fix signedness comparison warnings in Scene.cpp
Cast nodeList.size() to int32_t when comparing with mParentIndex to
eliminate -Wsign-compare warnings. Follows maintainer's preferred
pattern from issue #61.

Fixes #61

+++ Engine/Source/Engine/Assets/Scene.cpp
@@ -351,9 +351,9 @@ NodePtr Scene::Instantiate()
         for (uint32_t i = 0; i < mNodeDefs.size(); ++i)
         {
             NodePtr nodePtr;
-            NodePtr parent = (i > 0 && nodeList.size() > mNodeDefs[i].mParentIndex) ? nodeList[mNodeDefs[i].mParentIndex] : nullptr;
+            NodePtr parent = (i > 0 && int32_t(nodeList.size()) > mNodeDefs[i].mParentIndex) ? nodeList[mNodeDefs[i].mParentIndex] : nullptr;
 
-            if (i > 0 && nodeList.size() <= mNodeDefs[i].mParentIndex)
+            if (i > 0 && int32_t(nodeList.size()) <= mNodeDefs[i].mParentIndex)
             {
                 LogError("Out-of-order parent for node: %s", mNodeDefs[i].mName.c_str());
                 parent = nodeList[0];
